Add files via upload
This is a serverless web app, built with HTML, JavaScript and CSS. Its key features are that it has an instant update function (onSnapshot), uses an API (Firestore/Authentication). These were used to understand APIs, the onSnapshot function and the programming language CSS further as I am focusing on these sectors as of now.